+++ geminidr/core/primitives_spect.py
@@ -2671,6 +2671,10 @@ class Spect(PrimitivesBASE):
             write VAR (variance) plane?
         overwrite : bool
             overwrite existing files?
+        xunits: str
+            units of the x (wavelength/frequency) column
+        yunits: str
+            units of the data column
 
         Returns
         -------
@@ -2689,6 +2693,8 @@ class Spect(PrimitivesBASE):
         write_dq = params["dq"]
         write_var = params["var"]
         overwrite = params["overwrite"]
+        xunits = None if params["wave_units"] is None else u.Unit(params["wave_units"])
+        yunits = None if params["data_units"] is None else u.Unit(params["data_units"])
 
         for ad in adinputs:
             aperture_map = dict(zip(range(len(ad)), ad.hdr.get("APERTURE")))
@@ -2711,15 +2717,45 @@ class Spect(PrimitivesBASE):
                                 "1D array - continuing")
                     continue
 
+                output_frame = ext.wcs.output_frame
+                xdata = (ext.wcs(range(ext.data.size)) *
+                         (output_frame.unit[0] or u.nm))
+                if xunits is not None and xunits != xdata.unit:
+                    xdata = xdata.to(xunits)
                 data_unit = u.Unit(ext.hdr.get("BUNIT"))
-                t = Table((ext.wcs(range(ext.data.size)), ext.data),
+                ydata = ext.data * data_unit
+                equivalencies = u.spectral_density(xdata)
+                if yunits is not None:
+                    try:
+                        ydata = ydata.to(yunits, equivalencies=equivalencies)
+                    except u.core.UnitConversionError:
+                        try:
+                            ydata = (ydata / (ad.exposure_time() * u.s)).to(
+                                yunits, equivalencies=equivalencies)
+                        except u.core.UnitConversionError:
+                            log.warning(f"Cannot convert spectrum in {ad.filename}:"
+                                        f"{ext.hdr['EXTVER']} from {ydata.unit} to {yunits}")
+                            yunits = data_unit
+                else:
+                    yunits = data_unit
+
+                t = Table((xdata.value, ydata.value),
                           names=("wavelength", "data"),
-                          units=(ext.wcs.output_frame.unit[0], str(data_unit)))
+                          units=(xdata.unit, ydata.unit))
+                t.meta['comments'] = [f"Wavelength in {xdata.unit}, "
+                                      f"Data in {ydata.unit}"]
                 if write_dq:
                     t.add_column(ext.mask, name="dq")
                 if write_var:
-                    t.add_column(ext.variance, name="variance")
-                    t["variance"].unit = str(data_unit ** 2)
+                    stddev = np.sqrt(ext.variance) * data_unit
+                    try:
+                        stddev = stddev.to(yunits, equivalencies=equivalencies)
+                    except u.core.UnitConversionError:
+                        stddev = (stddev / (ad.exposure_time() * u.s)).to(
+                            yunits, equivalencies=equivalencies)
+                    var = stddev * stddev
+                    t.add_column(var.value, name="variance")
+                    t["variance"].unit = var.unit
                     var_col = len(t.colnames)
 
                 filename = (os.path.splitext(ad.filename)[0] +